+++ src/importProjectQuickStep.ts
@@ -4,7 +4,7 @@ import { ZephyrBoard } from "./ZephyrBoard";
 import { ZephyrProject } from "./ZephyrProject";
 import { ZephyrSDK } from "./ZephyrSDK";
 import { MultiStepInput } from "./utilities/MultiStepQuickPick";
-import { fileExists, getListZephyrSDKs, getListIARs, getSupportedBoards, getWestWorkspace, getWestWorkspaces, getZephyrSDK, normalizePath } from "./utils";
+import { fileExists, getListZephyrSDKs, getListIARs, getSupportedBoards, getWestWorkspace, getWestWorkspaces, validateProjectLocation, getZephyrSDK, normalizePath } from "./utils";
 
 export async function importProjectQuickStep(context: ExtensionContext) {
   const title = 'Import Project';
@@ -241,11 +241,6 @@ export async function importProjectQuickStep(context: ExtensionContext) {
     }
 	}
 
-  async function validateProjectLocation(location: string) {
-    if(!fileExists(location)) {
-      return "Invalid project path.";
-    }
-  }
 
   function shouldResume() {
     // Could show a notification with the option to resume.