+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "zephyr-workbench" extension will be documented in th
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.3.22]
+
+- Add import option in "Create new application" tab
+- Refractor IAR toolchain import messages
+
 ## [1.3.21]
 
 - Optimize setting to prevent conflicts with CMake Tools